move _server.js to helpers
