+++ package.json
@@ -69,9 +69,6 @@
     "xo": "*"
   },
   "xo": {
-    "esnext": true,
-    "ignores": [
-      "test/_*"
-    ]
+    "esnext": true
   }
 }
